Enforce python3 as interpreter for bc-show-eligible
Upon installing this in a new system, I was having issues to load some
modules (namely pygit2), and this was due to the fact that I only had
those installed for python3, and unfortunatelly my system default
interpreter is still the should-be-dead python2.

Making this change in my local installation fixed the issue for me.