CM-37347 - Add option --add-to-exclude-list
Currently `git bc-show-eligible` lists all non-cherry-picked commits you
have, even the ones which you will never port back (e.g. due to API
changes).

With the added option, you may exclude merge commits which you are no
longer interested in seeing.

To remove a hash from that list, you will have to remove the entry in
the file where it is saved (which is pointed out in the help text).
I assume that removing a commit from the exclude list is a rare event,
so I don't see harm in leaving this part manual.